version 2.0

+++ internal/model/message.go
@@ -2,5 +2,5 @@ package model
 
 type Message struct {
 	Text   string
-	ChatID int64
+	ChatID int
 }
